Update FocusTrapZone focus handling in ChoiceGroup
ChoiceGroup needs to implement special focus handling when used in a
FocusTrapZone to reimplement native browser behavior as detailed in 20859.

However, the change in 20859 introduced other issues when focusing ChoiceGroups.
Ultimately the issue with the ChoiceGroup behavior was that it was
inferring when FocusTrapZone was giving it focus and this inference was
not always correct.

This change makes the focus from FocusTrapZone explicit while testing for
it opt in so this change will not affect FocusTrapZone generally.

A data attribute is added to each of the bumpers in FocusTrapZone so
ChoiceGroup can easily test if it is receiving focus from FocusTrapZone.