update scrolling and measure code for DetailsList
This commit updates the scrollTop read and write code to properly
handle the case when _scollElement is `window`. In this case reading
scrollTop does not work, instead scrollY must be used. scrollY is
readonly, instead the window.scrollTo() method must be used. Internal
wrapper functions have been added to handle this behavior.

These functions also round all values up to the nearest integer via
Math.ceil(). Scrolling values may be fractional, particularly on systems
using display scaling (e.g., display is scaled 150%) and this ensures
we slightly over-estimate where the scroll position actually is.

Similar to the rounding of scroll position this commit also updates the
measurement functions to round x, y, width and height values up to the
nearest integer. This also provides a slight over-estimate on the height
of elements but eliminates componding error when measurements are
slightly smaller (e.g., 47.95 vs 48px).

Neither of these changes eliminates virtualization issue but the
scrolling change in particular minimizes it.