add mailreadcompose pane, fix some layout issues
The layout fixes feel a bit hacky but it seems like AutoSizer has issues
with CSS grid so I've instead done the main page layout with flexbox.

Ultimately I don't think this changes the utility of this app either
way so I'm fine with it though I like the cleanliness of grid.