stress test: adds test cases for v8, v9 and web components
Adds tests cases for

  - mounting
  - injecting styles
  - adding lots of nodes
  - removing all nodes
  - re-rendering the entire page

for Fluent V8, V9 and Web Components.

Introduces the concept of "renderers" that easily allow different
components to be slotting into tests with just a query parameter.

Moves RandomTree code to scripts to handle a ts-node compilation issue.
This isn't ideal but it doesn't seem worth spending lots of time figuring
the proper issue out right now.

Similarly, some of the randomization code has been moved around in scripts
because ts-node was having difficulty importing it.