read this commit for more details!
First pass on getting Why Did You Render working with Storybook.

It kinda works but it feels wobbly. WDYR does not work with the demo
react-storybook-addon app though it does work with SpinButton stories
and presumably any other stories since it's added as a decorator. This
code probably isn't using the SB API as effectively as possible but things
are to the point where WDYR can be toggled on and off via a SB addon.