revert DetailsRowField render order
Prior to v8.63.0 custom column render functions were invoked
before the parent wrapper JSX. This enabled users to modify `column`
props in the render function and have those changes reflected immediately
in the subsequent call to render the render function's parent JSX.

However, this is an anti-pattern as it is pushing prop up the React tree
and the render order was changed by PR #22081.

This commit keeps the changes from 22081 but moves the render order to
preserve the previous behavior. A unit test has been added to ensure
no future breaks will be made.